refactor: Find order on service delete order and adjust custom exceptions

+++ src/core/application/services/order_service.py
@@ -1,3 +1,6 @@
+from core.application.exceptions.commons_exceptions import (
+    NoDocumentsFoundException,
+)
 from core.application.services.order_number_service import OrderNumberService
 from core.domain.models.order import Order
 from core.domain.ports.repositories.order import OrderRepositoryInterface
@@ -31,5 +34,8 @@ class OrderService:
         return total_orders
 
     def delete_order(self, id: str) -> bool:
+        order = self.get_order_by_id(id)
+        if order is None:
+            raise NoDocumentsFoundException()
         was_order_deleted = self.repository.delete_order(id=id)
         return was_order_deleted
